Django/fukiappstore/fukiapp/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cloudinary==1.33.0
cryptography==41.0.3
Deprecated==1.2.14
Django==4.2.4
django-ckeditor==6.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-js-asset==2.1.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
jwcrypto==1.5.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pycparser==2.21
PyMySQL==1.1.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wrapt==1.15.0
